+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Desktop app lets you regain manual control by simply moving the mouse, pausing temporarily, and resuming.
 - Friendly error handling for different camera access failure scenarios.
 - API documentation.
+- `tracky-mouse.js` includes a CommonJS export, untested. I'm only testing script tag usage. I hope to switch to ES modules soon.
 
 ### Fixed
 - Function `average_points` was missing. It existed in JS Paint, the only place I tested the library, since I was extracting the code from JS Paint.